# Vectorize `basic_string::rfind` (the string needle overload)