Towards #5036.  

Similar to #5048, the differences:
 * The parameters calculations that use `_Start_at` are trickier
 * The `input_needle.empty()` special case in the tests is really special case this time
 * Used intentionally not-very-aligned strings, towards #5035

Benchmark                                   |     main  |      this   
--------------------------------------------|-----------|-------------
member_rfind<not_highly_aligned_string>/0   |   108 ns  |   23.2 ns   
member_rfind<not_highly_aligned_string>/1   |   101 ns  |   22.6 ns   
member_rfind<not_highly_aligned_string>/2   |  1298 ns  |    305 ns   
member_rfind<not_highly_aligned_string>/3   |  1554 ns  |    347 ns   
member_rfind<not_highly_aligned_string>/4   |  7914 ns  |   4870 ns   
member_rfind<not_highly_aligned_string>/5   | 15104 ns  |  15482 ns   
member_rfind<not_highly_aligned_wstring>/0  |  98.1 ns  |   39.7 ns   
member_rfind<not_highly_aligned_wstring>/1  |   101 ns  |   38.9 ns   
member_rfind<not_highly_aligned_wstring>/2  |  1288 ns  |    624 ns   
member_rfind<not_highly_aligned_wstring>/3  |  1559 ns  |    722 ns   
member_rfind<not_highly_aligned_wstring>/4  |  5645 ns  |   9777 ns   
member_rfind<not_highly_aligned_wstring>/5  | 15808 ns  |  20634 ns   

https://github.com/microsoft/STL/blob/a1bc1261795d4097cf7c12cfd0b5e2091809f281/benchmarks/src/search.cpp#L41-L49