@StephanTLavavej 2024-11-13
Results on my 5950X:

Benchmark                                    | Before   |    After | Speedup | Notes
---------------------------------------------|---------:|---------:|---------|------
`member_rfind<not_highly_aligned_string>/0`  |  84.4 ns |  21.5 ns | 3.93    | 0: Small, closer to end
`member_rfind<not_highly_aligned_string>/1`  |  78.9 ns |  21.1 ns | 3.74    | 1: Large, closer to end
`member_rfind<not_highly_aligned_string>/2`  |  1336 ns |   196 ns | 6.82    | 2: Small, closer to begin
`member_rfind<not_highly_aligned_string>/3`  |  1590 ns |   253 ns | 6.28    | 3: Large, closer to begin
`member_rfind<not_highly_aligned_string>/4`  | 10408 ns |  6720 ns | 1.55    | 4: Small, evil
`member_rfind<not_highly_aligned_string>/5`  | 12419 ns | 20189 ns | 0.62    | 5: Large, evil
`member_rfind<not_highly_aligned_wstring>/0` |   108 ns |  32.3 ns | 3.34    | 0: Small, closer to end
`member_rfind<not_highly_aligned_wstring>/1` |   104 ns |  31.5 ns | 3.30    | 1: Large, closer to end
`member_rfind<not_highly_aligned_wstring>/2` |  1845 ns |   441 ns | 4.18    | 2: Small, closer to begin
`member_rfind<not_highly_aligned_wstring>/3` |  2148 ns |   506 ns | 4.25    | 3: Large, closer to begin
`member_rfind<not_highly_aligned_wstring>/4` |  5756 ns | 13644 ns | 0.42    | 4: Small, evil
`member_rfind<not_highly_aligned_wstring>/5` | 15833 ns | 24849 ns | 0.64    | 5: Large, evil

These are great speedups with only the evil cases being pessimized, and they aren't absolutely horrible.