@StephanTLavavej 2024-11-14
:joy_cat: :rocket: :heart_eyes_cat: function! member `basic_string` important another vectorizing for Thanks